requirements.txt
asgiref==3.2.7
Django==3.0.5
django-braces==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
